solved bounce weirdness, and it sucked

+++ Centipede/Collision.cs
@@ -6,9 +6,9 @@ namespace Centipede
     public class Collision
     {
         public CharachterEnum entityType;
-        public int distance;
+        public float distance;
 
-        public Collision(CharachterEnum entityType, int distance)
+        public Collision(CharachterEnum entityType, float distance)
         {
             this.distance = distance;
             this.entityType = entityType;
